% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbubbleHeatmap.R
\name{ggbubbleHeatmap}
\alias{ggbubbleHeatmap}
\title{Draws a bubble heatmap}
\usage{
ggbubbleHeatmap(df)
}
\arguments{
\item{df}{Dataframe with the results of all GSEA analyses. Returned by
\code{\link[ggseabubble]{GSEAtable}}).}
}
\value{
A \code{ggplot2} object.
}
\description{
This function draws a bubble heatmap where tested pathways
appear in rows and comparisons in columns. The bubbles are colored by NES and
their size is proportional to the FDR.
}
