% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAtable.R
\name{GSEAtable}
\alias{GSEAtable}
\title{Creates a table from GSEA reports}
\usage{
GSEAtable(dir)
}
\arguments{
\item{dir}{Directory where the GSEA results can be found.}
}
\value{
A table with the results for all GSEA analyses.
}
\description{
This function creates a single table with all the GSEA results
present in a directory.
}
