% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbubbleHeatmap.R
\name{ggbubbleHeatmap}
\alias{ggbubbleHeatmap}
\title{Draws a bubble heatmap}
\usage{
ggbubbleHeatmap(
  df,
  n.perm = 1000,
  FDR.threshold = 0.05,
  bubble.size.range = c(1, 10),
  cluster.rows = TRUE,
  cluster.distance = "euclidean",
  cluster.method = "complete"
)
}
\arguments{
\item{df}{Dataframe with the results of all GSEA analyses. Returned by
\code{\link[ggseabubble]{GSEAtable}}.}

\item{n.perm}{Number of permutations used during the GSEA analysis.}

\item{FDR.threshold}{Threshold of significance. All bubbles with a FDR >=
\code{FDR.threshold} will have no fill in order to make the visualization
easier.}

\item{bubble.size.range}{Numeric vector of length 2 indicating the minimum
and maximum bubble size.}

\item{cluster.rows}{Logical. Whether rows should be clustered.}

\item{cluster.distance}{(\code{\link[stats]{dist}}'s \code{method}) Distance
measure to be used.}

\item{cluster.method}{(\code{\link[stats]{hclust}}'s \code{method})
Clustering method to be used.}
}
\value{
A \code{patchwork} object.
}
\description{
This function draws a bubble heatmap where tested pathways
appear in rows and comparisons in columns. The bubbles are colored by NES and
their size is proportional to the FDR.
}
